+++ RELEASE_NOTES.md
@@ -24,6 +24,7 @@
 * `PhysicsEngine` should no longer be instantiated in the Server `main.js` and in the client entry point.  Rather, it should be instantiated in the `GameEngine` subclass for your game.
 * `PhysicsEngine` constructor now does all initialization. Use of `init` function is deprecated.
 * `GameEngine` step method cannot be called without passing the `isReenact` argument.  Games which override the `step` method must pass this argument when calling the super method.
+* `GameObject` New `onRemoveFromWorld` to mirror `onAddToWorld` 
 * Objects are now instantiated with a reference to the gameEngine, and get and ID automatically
 * Method `isOwnedByPlayer` moved from `clientEngine` to `GameEngine`, and the `clientEngine` now sets the `playerId` in the `gameEngine`. `GameObject` constructor is therefore: constructor(gameEngine, options, props) and must call the super constructor correspondingly
 * The `GameWorld.getPlayerObject()` method has been removed, you can get the player objects using the `GameWorld.query()` method, passing a `playerId` attribute.